Added HTML white list
I renamed the preserveEmpty option to html, because this is mainly useful for HTML files. Every tag that is from type br, img, input, link and meta won't be output in the longer form when html option is specified. However the whitelist array might be obsolete, but I'll do further testing with it :)